added/updated tests
